refactor(api): translate models to its folder
We will implement detection models, thus we need to scope and
adjust the project structure inside models folder
